requirement.txt
fastapi
uvicorn
pytesseract
pillow